requirements.txt
numpy>=1.19.0
torch>=1.7.0
torchvision>=0.8.0
open3d>=0.13.0
opencv-python>=4.5.0
pyyaml>=5.4.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
tqdm>=4.50.0
tensorboard>=2.4.0
timm>=0.4.12
scipy>=1.6.0
plyfile>=0.7.0 